CR FIX: Move GenericAviationWeatherMessageScanner interface to from fi.fmi.converter.iwxxm to fi.fmi.converter.iwxxm.generic
