+++ src/main/java/fi/fmi/avi/converter/iwxxm/generic/GenericAviationWeatherMessageScanner.java
@@ -1,4 +1,4 @@
-package fi.fmi.avi.converter.iwxxm;
+package fi.fmi.avi.converter.iwxxm.generic;
 
 import javax.xml.xpath.XPath;
 import javax.xml.xpath.XPathExpressionException;
